Shit code pushing probably refactoring tomorrow

+++ src/cpu.h
@@ -9,7 +9,7 @@
 struct regis {
     uint8_t gregis[16];
     uint16_t I;
-    uint16_t VF;
+    uint8_t VF;
     uint8_t delayTimer;
     uint8_t soundTimer;
     uint16_t PC;
